requirements.txt
reportlab
PyYAML
Pillow
requests
